requirements.txt
openai
python_dotenv
langchain
langchain_openai
langchain-core
langchain-community
langgraph
